# docs/openapi.design.yaml
openapi: 3.0.3
info:
  title: OrbitPlay API — contrato de design
  version: 0.2.0-design
  description: |
    Contrato **de design** do backend do OrbitPlay, derivado de
    `docs/BACKEND-SPEC.md` (handoff v1.6). Serve para documentar e começar o
    desenvolvimento antes da implementação existir.

    ## Relação com `openapi.json`

    `openapi.json` na raiz do repositório é **gerado** a partir dos decorators
    do Nest (`pnpm openapi:generate`) e descreve apenas o que já está
    implementado. Este arquivo descreve o **alvo**. Conforme cada endpoint é
    implementado, ele passa a existir no `openapi.json` gerado — que continua
    sendo a fonte da verdade do que roda. Este YAML nunca é gerado; é editado à
    mão e vale como especificação.

    Endpoints já implementados estão marcados com `x-status: implemented`; os
    demais, com `x-status: todo`.

    ## Escopo

    Esta fase cobre a plataforma. Estão **fora**: cobrança do estúdio, carteira
    e saque do jogador, impulsionamento do feed, Orbit Plug-in e ingestão de
    telemetria, insights de IA e transcrição (ASR). Ver seção 10 da
    especificação.

    ## Convenções

    - **Tenancy:** toda rota autenticada é escopada pela organização do token.
      Recurso de outra organização responde `404`, nunca `403`, para não vazar
      existência.
    - **Papel:** vem da membership ativa carregada no access token, nunca do
      corpo da requisição.
    - **Erros:** envelope único `ErrorEnvelope` em toda falha.
...
servers:
...
security:
  - bearer: []
...
paths:
...
  /orgs/current:
    get:
      tags: [orgs]
      summary: Organização da sessão
      x-status: implemented
      responses:
        '200':
          description: Organização atual
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Org' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [orgs]
      summary: Atualizar dados da organização
      description: 'Papéis: `owner`, `admin`.'
      x-status: todo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrgRequest' }
      responses:
        '200':
          description: Organização atualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Org' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /orgs/members:
    get:
      tags: [orgs]
      summary: Listar membros
      description: |
        Tela 20 exige busca e filtro. Hoje implementado sem paginação nem
        filtros — estender.
      x-status: partial
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          description: Busca por nome ou e-mail
          schema: { type: string }
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/Role' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/MembershipStatus' }
      responses:
        '200':
          description: Página de membros
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberList' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /orgs/members/invite:
    post:
      tags: [orgs]
      summary: Convidar membro
      description: |
        Cria membership com status `invited` e dispara e-mail de convite. Não
        cria senha — o convidado define a dele pelo fluxo de redefinição
        (RN-04, Tela 20).
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InviteMemberRequest' }
      responses:
        '201':
          description: Convite criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Usuário já é membro da organização
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /orgs/members/{userId}/role:
    patch:
      tags: [orgs]
      summary: Alterar papel de um membro
      description: |
        RN-02 (Tela 20): ação crítica, exige `confirm: true` explícito.

        RN-03: rebaixar o **último Owner ativo** é recusado com `409`.

        RN-05: gera registro em `audit_log` com autor, data e ação.

        Papel exigido: `owner`. A RN-01 admite "Admin com permissão
        específica", mas permissões por usuário não existem no projeto —
        ver DECISIONS.md §3.

        O `409` de conflito de edição concorrente **não** está implementado:
        `ChangeRoleRequest` não carrega versão nem `updatedAt`, então não há
        o que comparar. Vale também para `ChangeStatusRequest`.
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeRoleRequest' }
      responses:
        '200':
          description: Papel alterado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Último Owner ativo, ou conflito de edição concorrente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /orgs/members/{userId}/status:
    patch:
      tags: [orgs]
      summary: Ativar ou desativar um membro
      description: |
        RN-06 (Tela 20): desativação lógica; não há exclusão física quando há
        histórico de testes associado. Desativar o último Owner ativo → `409`.
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeStatusRequest' }
      responses:
        '200':
          description: Status alterado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Último Owner ativo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /orgs/members/{userId}/password-reset:
    post:
      tags: [orgs]
      summary: Disparar redefinição de senha para um membro
      description: |
        RN-04 (Tela 20): o administrador **nunca** define nem vê a senha de
        outro usuário; apenas dispara o fluxo de redefinição.
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '202':
          description: E-mail de redefinição enviado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orgs/members/{userId}:
    delete:
      tags: [orgs]
      summary: Remover membro
      description: |
        RN-06: desativação lógica quando há histórico associado. Papel `owner`.
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204': { description: Membro removido ou desativado }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Último Owner ativo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /games:
    get:
      tags: [games]
      summary: Listar jogos do estúdio
      description: |
        RN-03 (Tela 03): as métricas do card são **agregadas no backend** a
        partir dos testes do jogo. Filtros `q`/`status` e paginação por cursor.
        `sessionsValid`, `playersTotal` e `averageRating` ficam zerados até
        M8/M13 gravarem linhas.
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/GameStatus' }
      responses:
        '200':
          description: Página de jogos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [games]
      summary: Criar jogo
      description: |
        RN-02 (Tela 04): o jogo é sempre criado dentro da organização da sessão;
        `organizationId` não é aceito no corpo.
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGameRequest' }
      responses:
        '201':
          description: Jogo criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Game' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Slug já usado nesta organização
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /games/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [games]
      summary: Obter jogo
      x-status: implemented
      responses:
        '200':
          description: Jogo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Game' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [games]
      summary: Atualizar jogo
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateGameRequest' }
      responses:
        '200':
          description: Jogo atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Game' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [games]
      summary: Remover jogo
      description: Exclusão lógica (`deleted_at`).
      x-status: implemented
      responses:
        '204': { description: Jogo removido }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /games/{id}/assets/upload-url:
    post:
      tags: [games]
      summary: Obter URL assinada para upload de imagem
      description: |
        RN-05 (Tela 04): formato, tamanho e falha de envio são validados **no
        backend**. Esta rota valida `contentType` e `sizeBytes` antes de assinar;
        a confirmação em `POST /games/{id}/assets` valida o objeto de fato
        gravado.

        O binário vai direto do cliente para o storage — a API nunca faz proxy.
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetUploadUrlRequest' }
      responses:
        '201':
          description: URL de upload assinada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadUrlResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /games/{id}/assets:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [games]
      summary: Confirmar upload e vincular imagem ao jogo
      description: |
        Verifica a existência do objeto no storage antes de gravar a linha. Um
        upload que não chegou não vira asset.
      x-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmAssetRequest' }
      responses:
        '201':
          description: Asset vinculado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameAsset' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: Objeto ausente no storage, formato ou tamanho inválido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /games/{id}/assets/{assetId}:
    delete:
      tags: [games]
      summary: Remover imagem do jogo
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Asset removido }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /games/{id}/summary:
    get:
      tags: [games]
      summary: Resumo do jogo para a tela de detalhes
      description: 'Tela 05. Banner, disponibilidade e indicadores agregados.'
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Resumo do jogo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameSummary' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /games/{id}/specs:
    get:
      tags: [games]
      summary: Especificações técnicas do jogo
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Especificações
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameSpecs' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------- M4 test-models
...
  /player/games/{gameId}:
    get:
      tags: [player-feed]
      summary: Detalhes de um jogo para o jogador
      description: 'Tela 15. Banner, disponibilidade, jogadores e testes abertos.'
      x-status: todo
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Jogo na visão do jogador
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlayerGame' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token de curta duração devolvido por `/auth/login`.
    cookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: 'Refresh token httpOnly, rotacionado a cada `/auth/refresh`.'
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    GameId:
      name: gameId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      description: Itens por página.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      description: Cursor opaco devolvido como `nextCursor` na página anterior.
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Ausente, inválido ou expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Papel sem permissão para esta ação
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: |
        Recurso inexistente — ou pertencente a outra organização. As duas
        situações respondem igual, de propósito, para não vazar existência.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Dados inválidos, com erro por campo em `fieldErrors`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      description: |
        Envelope único de erro da API. Serializado exclusivamente pelo
        `HttpExceptionFilter` — nenhum controller monta erro à mão.
      required: [statusCode, code, message, requestId]
      properties:
        statusCode: { type: integer, example: 422 }
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, example: Dados inválidos }
        fieldErrors:
          type: object
          description: Mensagem por campo, presente em falhas de validação.
          additionalProperties: { type: string }
          example: { title: Título obrigatório }
        requestId:
          type: string
          description: Correlaciona a resposta com os logs estruturados.
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - TOO_MANY_REQUESTS
        - UNPROCESSABLE_ENTITY
        - INTERNAL_ERROR
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    Role:
      type: string
      enum: [owner, admin, studio, player]
    MembershipStatus:
      type: string
      enum: [active, invited, disabled]
    GameStatus:
      type: string
      enum: [draft, active, archived]
...
    Platform:
      type: string
      enum: [windows, macos, linux, android, ios, web]
...
    UploadUrlResponse:
      type: object
      description: |
        URL assinada para o cliente enviar bytes **direto ao storage**. A API
        nunca faz proxy de binário.
      required: [uploadUrl, storageKey, expiresAt]
      properties:
        uploadUrl: { type: string, format: uri }
        storageKey: { type: string }
        expiresAt: { type: string, format: date-time }
        maxSizeBytes: { type: integer }

    # ---- auth ----
...
    Org:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
    UpdateOrgRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
    Member:
      type: object
      required: [userId, email, displayName, role, status]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        status: { $ref: '#/components/schemas/MembershipStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: Usado para detectar edição concorrente.
    MemberList:
      type: object
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Member' }
        nextCursor: { type: string, nullable: true }
    InviteMemberRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
    ChangeRoleRequest:
      type: object
      required: [role, confirm]
      properties:
        role: { $ref: '#/components/schemas/Role' }
        confirm:
          type: boolean
          description: 'RN-02: ação crítica, exige confirmação explícita.'
          enum: [true]
    ChangeStatusRequest:
      type: object
      required: [status, confirm]
      properties:
        status: { $ref: '#/components/schemas/MembershipStatus' }
        confirm: { type: boolean, enum: [true] }
...
    Game:
      type: object
      required: [id, organizationId, title, slug, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        genre: { type: string, nullable: true }
        platform: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/GameStatus' }
        coverUrl: { type: string, format: uri, nullable: true }
        bannerUrl: { type: string, format: uri, nullable: true }
        metrics: { $ref: '#/components/schemas/GameMetrics' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    GameMetrics:
      type: object
      description: |
        RN-03 (Tela 03): agregados a partir dos testes do jogo, calculados no
        backend.
      properties:
        testsTotal: { type: integer }
        testsActive: { type: integer }
        sessionsValid: { type: integer }
        playersTotal: { type: integer }
        averageRating: { type: number, format: float, nullable: true }
    GameList:
      type: object
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Game' }
        nextCursor: { type: string, nullable: true }
    CreateGameRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        slug:
          type: string
          description: Derivado do título quando omitido.
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description: { type: string, maxLength: 5000 }
        genre: { type: string, maxLength: 100 }
        platform: { type: string, maxLength: 100 }
        status: { $ref: '#/components/schemas/GameStatus' }
    UpdateGameRequest:
      type: object
      description: Todos os campos opcionais.
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        description: { type: string, maxLength: 5000 }
        genre: { type: string, maxLength: 100 }
        platform: { type: string, maxLength: 100 }
        status: { $ref: '#/components/schemas/GameStatus' }
    AssetKind:
      type: string
      enum: [cover, banner, screenshot]
    AssetUploadUrlRequest:
      type: object
      required: [kind, contentType, sizeBytes, fileName]
      properties:
        kind: { $ref: '#/components/schemas/AssetKind' }
        contentType: { type: string, enum: ['image/png', 'image/jpeg', 'image/webp'] }
        sizeBytes: { type: integer, minimum: 1 }
        fileName: { type: string }
    ConfirmAssetRequest:
      type: object
      required: [kind, storageKey]
      properties:
        kind: { $ref: '#/components/schemas/AssetKind' }
        storageKey: { type: string }
    GameAsset:
      type: object
      required: [id, kind, url, createdAt]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/AssetKind' }
        url: { type: string, format: uri }
        contentType: { type: string, nullable: true }
        sizeBytes: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
    GameSummary:
      type: object
      description: 'Tela 05 — cabeçalho e indicadores do jogo.'
      required: [game, availability, metrics]
      properties:
        game: { $ref: '#/components/schemas/Game' }
        availability:
          type: string
          enum: [available, unavailable]
        metrics: { $ref: '#/components/schemas/GameMetrics' }
        canEdit:
          type: boolean
          description: 'RN-03 (Tela 05): reflete a permissão real do usuário.'
    GameSpecs:
      type: object
      properties:
        minimumRequirements: { type: object, additionalProperties: { type: string } }
        recommendedRequirements: { type: object, additionalProperties: { type: string } }
        supportedPlatforms:
          type: array
          items: { $ref: '#/components/schemas/Platform' }
        languages:
          type: array
          items: { type: string }

    # ---- test models ----
...
    PlayerGame:
      type: object
      required: [game, openTests]
      properties:
        game: { $ref: '#/components/schemas/Game' }
        playersTotal: { type: integer }
        openTests: { type: integer }
        maxRewardCents: { type: integer, nullable: true }
        specs: { $ref: '#/components/schemas/GameSpecs' }
...